requirements.txt
Django==4.2.3
Pillow>=10.4.0
jinja2>=3.1
gunicorn==20.1.0
typing-extensions>=4.7.0
aiohttp>=3.9.0
click>=8.1.0
wheel>=0.42.0
multidict>=6.0
whitenoise==6.5.0
reportlab>=4.4.10
dj-database-url==2.1.0
python-dotenv==1.0.0
qrcode==7.4.2
djangorestframework==3.14.0
django-ratelimit==4.1.0
django-redis==5.4.0
celery==5.3.6
redis==5.0.1
drf-spectacular==0.26.5
django-allauth==0.54.0
psycopg2-binary==2.9.9
